Move year_rollover_helper
